fixed routing inssues with packages
