+++ client/package.json
@@ -12,6 +12,7 @@
     "@types/react-dom": "^18.3.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-router-dom": "^6.27.0",
     "react-player": "^2.16.0",
     "react-scripts": "5.0.1",
     "typescript": "^4.9.5",
