+++ client/src/App.css
@@ -36,3 +36,7 @@
     transform: rotate(360deg);
   }
 }
+body {
+  /* background-color: #8EC5FC; */
+  background-image: linear-gradient(62deg, #8EC5FC 0%, #c3fcec 100%);
+}
