+++ client/src/App.tsx
@@ -3,20 +3,19 @@ import logo from './logo.svg';
 import './App.css';
 import Navbar from "./component/navbar/Navbar"
 import { ToggleThemeButton } from './component/toggleTheme/ToggleThemeButton';
-import { VideoPlayer } from './views/videoView';
-import { Route, Routes } from 'react-router-dom';
+import { VideoPlayer } from './component/VideoPlayer';
+import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import { ToggleMusicButton } from './component/toggleMusic/ToggleMusicButton';
+import VideoPlayerView from './views/VideoPlayerView';
 
 function App() {
   return (
     <div>
-      <Navbar />
-      {/* <Routes>
-        <Route path="/" element={<VideoPlayer/>} />
-      </Routes> */}
-      <VideoPlayer/>
-      <ToggleMusicButton />
-      <ToggleThemeButton />
+      <BrowserRouter>
+        <Routes>
+          <Route path = "/" element = {<VideoPlayerView/>} />
+        </Routes>
+      </BrowserRouter>
     </div>
 
   );
