+++ client/src/component/VideoPlayer.css
@@ -1,0 +1,22 @@
+.video-player-container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    width: 80%;
+    margin: 100px  auto; 
+  }
+  
+  .video-element {
+    width: 100%;
+    max-width: 800px;
+    border-radius: 8px;
+    margin-bottom: 10px;
+  }
+  
+  .controls {
+    display: flex;
+    gap: 10px;
+  }
+  
+
+
