+++ client/src/component/VideoPlayer.tsx
@@ -1,0 +1,10 @@
+import ReactPlayer from 'react-player'
+import './VideoPlayer.css';
+
+export const VideoPlayer: React.FC = () => {
+
+    return(<div className='video-player-container'>
+        <ReactPlayer url='https://www.youtube.com/watch?v=1ZYbU82GVz4' />
+        </div>) 
+};
+
