+++ client/src/component/toggleTheme/ToggleThemeButton.tsx
@@ -1,10 +1,9 @@
-import { useContext, useState } from "react";
+import { useState } from "react";
 import { FaMoon } from "react-icons/fa";
 import { ThemeContext, themes } from "../../context/ThemeContext";
 import "../ToggleButton.css";
 
 export function ToggleThemeButton( ){
-    // const theme = useContext(ThemeContext);
     const [currentTheme, setCurrentTheme] = useState(themes.light);
     const toggleTheme = () => {
       setCurrentTheme(currentTheme === themes.light ? themes.dark : themes.light);
