+++ client/src/context/ThemeContext.ts
@@ -14,3 +14,8 @@ export const themes = {
 };
 
 export const ThemeContext = React.createContext(themes.light);
+
+
+
+
+
